requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
certifi==2025.4.26
colorama==0.4.6
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==4.2.10
django-crispy-forms==2.0
django-filter==23.2
django-import-export==3.2.0
django-model-utils==5.0.0
django-notifications-hq==1.8.2
django-widget-tweaks==1.4.12
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
jsonfield==3.1.0
numpy==2.2.6
odfpy==1.4.1
openai==1.0.0
openpyxl==3.1.5
Pillow==10.0.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.2
reportlab==4.0.4
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
tablib==3.8.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
